squash: move checkIsArrayBufferView to util
This commit moves checkIsArrayBufferView to crypto/util to allow it to
be used by other modules in crypto.
