+++ lib/internal/crypto/sig.js
@@ -14,10 +14,10 @@ const {
   RSA_PKCS1_PADDING
 } = process.binding('constants').crypto;
 const {
+  checkIsArrayBufferView,
   getDefaultEncoding,
   toBuf
 } = require('internal/crypto/util');
-const { isArrayBufferView } = require('internal/util/types');
 const { Writable } = require('stream');
 const { inherits } = require('util');
 
@@ -135,14 +135,3 @@ module.exports = {
   Sign,
   Verify
 };
-
-function checkIsArrayBufferView(name, buffer) {
-  if (!isArrayBufferView(buffer)) {
-    throw new ERR_INVALID_ARG_TYPE(
-      name,
-      ['string', 'Buffer', 'TypedArray', 'DataView'],
-      buffer
-    );
-  }
-  return buffer;
-}
