+++ lib/internal/crypto/util.js
@@ -83,7 +83,19 @@ function timingSafeEqual(buf1, buf2) {
   return _timingSafeEqual(buf1, buf2);
 }
 
+function checkIsArrayBufferView(name, buffer) {
+  if (!isArrayBufferView(buffer)) {
+    throw new ERR_INVALID_ARG_TYPE(
+      name,
+      ['string', 'Buffer', 'TypedArray', 'DataView'],
+      buffer
+    );
+  }
+  return buffer;
+}
+
 module.exports = {
+  checkIsArrayBufferView,
   getCiphers,
   getCurves,
   getDefaultEncoding,
